These are the initial classes for the new image annotation functionality in Phenote.  This includes a new Menu to open images into their own panel; a GUIComponent that houses the image, any ROIs drawn on top of it, and a toolbar (to be added in the future).  The ROIs are just a dummy set of Shapes...to be defined by the user in the future.
This functionality will remain invisible until the menu is added to the startup task.

The initial version of these classes only allows for a single image to be viewed at a time (the ImageManager will need to be modified to accommodate multiple images.)

The ImageInfo Panel doesn't show anything yet.