* Fixed sclera def.    * spatial disjointness between pec and pel complexes    * Fixed egregious upper arm bone part_of hindlimb error. Also added spatial disjointness axiom, impossible for this to happen again    * add symm axiom
! uberon definition changes
!
! Files used:
! file 1 (old): date unknown, cvs revision unknown
! file 2 (new): date unknown, cvs revision unknown

!
! Format of first line of each entry:
! Term ID : term name  ( total # annots / # direct annots)


UBERON:0001773 : sclera   
OLD: The sclera, also known as the white part of the eye, is the opaque (usually white, though certain animals, such as horses and lizards, can have black sclera), fibrous, protective, outer layer of the eye containing collagen and elastic fiber. It is derived from the neural crest. In children, it is thinner and shows some of the underlying pigment, appearing slightly blue. In the elderly, however, fatty deposits on the sclera can make it appear slightly yellow. The sclera forms the posterior five sixths of the connective tissue coat of the globe. It is continuous with the dura mater and the cornea, and maintains the shape of the globe, offering resistance to internal and external forces, and provides an attachment for the extraocular muscle insertions. The sclera is perforated by plenty of nerves and vessels passing through the posterior scleral foramen, the hole that is formed by the optic nerve. At the optic disk the outer two-thirds of the sclera continues with the dura mater (outer coat of the brain) via the dural sheath of the optic nerve. The inner third joins with some choroidal tissue to form a plate (lamina cribrosa) across the optic nerve with perforations through which the optic fibers (fasciculi). The thickness of the sclera varies from 1mm at the posterior pole to 0.3 mm just behind the rectus muscle insertions. The sclera's blood vessels are mainly on the surface, and together with the conjunctiva (which lies on top) This is a thin layer covering the sclera. Along with the vessels of the conjunctiva, those of the sclera renders the inflamed eye bright red. [WP,unvetted].
NEW: Opaque fibrous outer layer of the eyeball[ZFA,Kardong,WP].



uberon
Ontology Comparison Report
==========================

Files used:
- file 1 (old): uberon-lastbuild.obo; data version: 2013-02-04
- file 2 (new): uberon.obo; data version: 2013-02-06

NEW TERMS
=========

None


OBSOLETIONS
===========

None


TERM MERGES
===========

None


TERM CHANGES
============

KEY:  'A' : added, 'D' : deleted, 'C' : changed

Term ID	def	synonym	is_a	relationship	
UBERON:0000976			1 D		
UBERON:0001773	C				
UBERON:0002102					
UBERON:0004109			1 D, 1 A		
UBERON:0004250			2 D, 1 A	1 D, 1 A	
UBERON:0005928		1 D, 1 A			
UBERON:0006245			1 D		
UBERON:0006246			1 D		
UBERON:0010709					



OTHER STANZA TYPES
==================

Stanzas Edited
==============

Stanza type	ID	Name
Typedef	adjacent_to	adjacent_to



FILE STATS
==========

Term counts for the old and new files

Ontology	Old file	New file	Change
uberon	8275	8275	
	(6724 defined)	(6724 defined)	

Obsolete	196	196	0
	(2.3% of total)	 (2.3% of total)	(0.0%)

Totals	8275	8275	0 (0.0%)
	6724 (81.3%) defined	6724 (81.3%) defined	
	8471 inc. obs	8471 inc. obs